+++ src/greynoise/api/__init__.py
@@ -287,8 +287,10 @@ class GreyNoise(object):
         :rtype: dict
 
         """
-        if isinstance(ip_addresses, str):
+        if isinstance(ip_addresses, str) and "," not in ip_addresses:
             ip_addresses = [ip_addresses]
+        elif isinstance(ip_addresses, str) and "," in ip_addresses:
+            ip_addresses = ip_addresses.split(",")
 
         LOGGER.debug("Getting noise status...", ip_addresses=ip_addresses)
         ip_addresses = [