+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Version `dev`_
 
   * Updated json_formatter for query commands to return data only as New Line Delimited JSON
 
+* API client:
+
   * Updated logic in quick to better handle non-list format input ('ip_1,ip_2')instead of
     ['ip_1','ip_2']