+++ CHANGELOG.rst
@@ -6,13 +6,16 @@ Version `dev`_
 ================
 **Date**: unreleased
 
-* API client:
+* CLI:
+
+  * Updated json_formatter for query commands to return data only as New Line Delimited JSON
 
   * Add "include_invalid" option to QUICK lookup to return invalid IPs as part of the JSON response
 
 * Dependencies:
 
   * Updated sphinx to 3.4.0
+  * Updated structlog to 20.2.0 for python 3.6 and 3.7
 
 Version `0.6.0`_
 ================