+++ CHANGELOG.rst
@@ -6,11 +6,16 @@ Version `dev`_
 ================
 **Date**: unreleased
 
-* CLI:
-
-  * Updated json_formatter for query commands to return data only as New Line Delimited JSON
+* API client:
 
   * Add "include_invalid" option to QUICK lookup to return invalid IPs as part of the JSON response
+
+  * Added support for new /riot endpoint
+
+* CLI:
+
+  * Added support for new riot command
+  * Updated json_formatter for query commands to return data only as New Line Delimited JSON
 
 * Dependencies: