+++ CHANGELOG.rst
@@ -9,8 +9,9 @@ Version `dev`_
 * API client:
 
   * Add "include_invalid" option to QUICK lookup to return invalid IPs as part of the JSON response
-
   * Added support for new /riot endpoint
+  * Updated logic in quick to better handle non-list format input ('ip_1,ip_2')instead of
+    ['ip_1','ip_2']
 
 * CLI: