+++ CHANGELOG.rst
@@ -8,6 +8,16 @@ Version `dev`_
 
 * API client:
 
+  * Add "include_invalid" option to QUICK lookup to return invalid IPs as part of the JSON response
+  * Added support for new /riot endpoint
+  * Updated logic in quick to better handle non-list format input ('ip_1,ip_2')instead of
+    ['ip_1','ip_2']
+
+* CLI:
+
+  * Added support for new riot command
+  * Updated json_formatter for query commands to return data only as New Line Delimited JSON
+
 * Dependencies:
 
   * Updated sphinx to 3.4.0