+++ CHANGELOG.rst
@@ -2,9 +2,9 @@
 Changelog
 =========
 
-Version `dev`_
+Version `0.7.0`_
 ================
-**Date**: unreleased
+**Date**: January 07, 2021
 
 * API client:
 
@@ -140,4 +140,4 @@ Version `0.2.0`_
 .. _`0.4.1`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.4.0...0.4.1
 .. _`0.5.0`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.4.1...0.5.0
 .. _`0.6.0`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.5.0...0.6.0
-.. _`dev`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.6.0...HEAD
+.. _`0.7.0`: https://github.com/GreyNoise-Intelligence/pygreynoise/compare/v0.6.0...0.7.0